Split settings panels logic and add support for custom panels
